feat: Switch to pnpm
