+++ package.json
@@ -22,21 +22,28 @@
     "ember-addon-migrator": "./bin.js"
   },
   "exports": {
-    "import": "./index.js"
+    ".": {
+      "import": "./index.js"
+    },
+    "./captured-addon": {
+      "import": "./src/captured-addon/index.js"
+    },
+    "./captured-addon/utils/fs-helpers": {
+      "import": "./src/captured-addon/utils/fs-helpers.js"
+    }
   },
   "files": [
     "src/**/*",
     "bin.js"
   ],
   "scripts": {
-    "publish:dry-run": "yarn publish:release --dry-run",
-    "publish:release": "yarn semantic-release",
+    "publish:dry-run": "pnpm publish:release --dry-run",
+    "publish:release": "pnpm semantic-release",
     "build:types": "tsc --build",
-    "lint": "yarn lint:js",
+    "lint": "pnpm lint:js",
     "lint:js": "eslint .",
-    "lint:fix": "yarn lint:js --fix",
-    "test": "vitest --coverage --no-watch",
-    "test:watch": "vitest --watch"
+    "lint:fix": "pnpm lint:js --fix",
+    "test": "pnpm --filter 'tests' test"
   },
   "dependencies": {
     "chalk": "^5.0.0",
@@ -67,7 +74,6 @@
     "adm-zip": "^0.5.9",
     "babel-eslint": "^10.1.0",
     "broccoli-test-helper": "^2.0.0",
-    "c8": "^7.11.0",
     "eslint": "^7.0.0",
     "eslint-plugin-decorator-position": "^4.0.1",
     "eslint-plugin-import": "^2.25.4",
@@ -76,12 +82,9 @@
     "lodash": "^4.17.21",
     "node-fetch": "^3.2.0",
     "prettier": "^2.5.1",
-    "scenario-tester": "^2.0.1",
     "semantic-release": "^19.0.0",
     "types-package-json": "^2.0.39",
-    "typescript": "4.5.4",
-    "vite": "^2.7.13",
-    "vitest": "^0.3.0"
+    "typescript": "4.5.4"
   },
   "engines": {
     "node": ">=16.0.0"
