+++ pnpm-workspace.yaml
@@ -1,0 +1,7 @@
+packages:
+  # allows the root of the repo to be its own workspace
+  - "."
+  # the tests folder is own package.
+  # this way, all the test dependencies are separate from the
+  # main package.json
+  - "./tests"
