+++ tests/index.test.ts
@@ -1,6 +1,6 @@
 import { describe, test } from 'vitest';
 
-import { fastVerify, sampleAddons, verify } from './helpers';
+import { verify } from './helpers';
 
 describe('basic conversion', () => {
   test('js addon', async () => {
@@ -13,15 +13,5 @@ describe('basic conversion', () => {
 
   test('sample addon [ember-addon-output_4.2.0]', async () => {
     await verify('ember-addon-output_4.2.0');
-  });
-});
-
-describe('snapshots conversion', () => {
-  const addons = sampleAddons();
-
-  addons.forEach((addonName) => {
-    test(`sample addon [${addonName}]`, async () => {
-      await fastVerify(addonName);
-    });
   });
 });
