+++ tests/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "tests",
+  "private": true,
+  "scripts": {
+    "test": "vitest --coverage --no-watch",
+    "test:watch": "vitest --watch"
+  },
+  "devDependencies": {
+    "@vitest/coverage-c8": "^0.23.2",
+    "broccoli-test-helper": "^2.0.0",
+    "c8": "^7.11.0",
+    "ember-addon-migrator": "workspace:..",
+    "node-fetch": "^3.2.0",
+    "scenario-tester": "^2.0.1",
+    "tinypool": "^0.3.0",
+    "typescript": "4.8.3",
+    "vite": "^3.1.1",
+    "vitest": "^0.23.2"
+  },
+  "engines": {
+    "node": ">=16.0.0"
+  },
+  "volta": {
+    "extends": "../package.json"
+  }
+}
